[BAAAS-40] Created first version of domain model to support demo 2.0 API
Created the database schema to support most functionality of the BAaaS API for demo 2.0.
Created a central manager class that handles the majority of the lifecycle management. API controllers should be able to delegate to this.
Tests of all of the above to verify the majority of the lifecycle use-cases